refactor(rooms): extract item movement
